chore(CU-2p1bn1p): add meta tags in missing pages, refactor some files
