+++ apps/web/src/pages/About/About.tsx
@@ -1,0 +1,14 @@
+import { NextSeo } from 'next-seo';
+
+import Members from '@/components/Members';
+
+import { StyledAbout } from './styles';
+
+const About = () => (
+  <StyledAbout>
+    <Members />
+    <NextSeo title='About Project Shelf' />
+  </StyledAbout>
+);
+
+export default About;
